+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Re add channel 7 for HmIPW-WRC6
 - Reduce log level for exceptions in fetch_paramset_description
+- Filter SSLErrors by code
 
 # Version 2023.9.0 (2023-09-03)