+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2023.9.1 (2023-09-06)
 
 - Re add channel 7 for HmIPW-WRC6
+- Reduce log level for exceptions in fetch_paramset_description
+- Filter SSLErrors by code
 
 # Version 2023.9.0 (2023-09-03)